Introduces 'state' object in class-based components
props and state  are CORE concepts of React. Actually, only changes in props  and/ or state  trigger React to re-render your components and potentially update the DOM in the browser

'props' allow you to pass data from a parent (wrapping) component to a child (embedded) component.

'state' is used to change the component, well, state from within.